% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfresco.R
\name{alf_defaults}
\alias{alf_defaults}
\title{Common default values}
\usage{
alf_defaults()
}
\value{
a named list.
}
\description{
Common default values in the context of ALFRESCO modeling.
}
\details{
This function returns a list of default values like standard file paths on the SNAP Atlas cluster
for running the model and processing outputs and paths to standard input maps.
}
\examples{
alf_defaults()$atlas_shiny_dir
}
