% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{alf_dist_inputs}
\alias{alf_dist_inputs}
\title{Table inputs from extracted data}
\usage{
alf_dist_inputs(project, in_dir)
}
\arguments{
\item{project}{character, see details.}

\item{in_dir}{path to data extractions, the parent directory with the three subdirectories, \code{fsv}, \code{veg} and \code{age}. See details.}
}
\value{
a data frame.
}
\description{
Create a data frame of inputs from extraction .rds files for probability distribution estimation.
}
\details{
This function creates a data frame containing columns of metadata based on the .rds files
stored in the \code{fsv}, \code{veg} and \code{age} extraction subdirectories.
These files are output during ALFRESCO data extraction by \link{run_alf_extraction}.
The columns of the data frame describe the random variable set (\code{fsv}, \code{veg} or \code{age}),
location group (region set), specific location (region),
model (a GCM or CRU), RCP (including "historical"), and the fire management options (FMO) treatment.

If \code{in_dir} is not provided, a default is taken from \code{alfdef()$alf_extract_dir} and \code{project}, so \code{project}
must be provided. \code{project} is ignored if the full project path is provided via \code{in_dir}.
Provide a directory or provide a project name and rely on the default extractions directory.
}
\examples{
\dontrun{alf_dist_inputs("JFSP")}
}
