% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{alf_extract_rscript}
\alias{alf_extract_rscript}
\title{R script for ALFRESCO data extraction}
\usage{
alf_extract_rscript(out_dir = alfdef()$alf_slurm_dir,
  file = "alf_extract.R")
}
\arguments{
\item{out_dir}{output directory of R script.}

\item{file}{R script.}
}
\description{
Generate a template R script on disk to be run by a slurm script made using \link{alf_extract_slurm}.
This function does not need to be called directly if \link{alf_extract_slurm} is run with the default \code{copy_rscript = TRUE}.
}
\details{
This function is intended to be used in conjunction with \link{alf_extract_slurm}.
It creates a copy on disk at \code{out_dir} of a template R script contained in the \code{alfresco} package.
}
\examples{
\dontrun{alf_extract_rscript()}
}
