% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{alf_prep_rscript}
\alias{alf_prep_rscript}
\title{R script for ALFRESCO probability distributions}
\usage{
alf_prep_rscript(out_dir = alfdef()$alf_slurm_dir, file = "alf_prep.R")
}
\arguments{
\item{out_dir}{output directory of R script.}

\item{file}{R script.}
}
\description{
Generate a template R script for curating probability distributions from extracted ALFRESCO data.
}
\details{
Create a template R script on disk to be run by a slurm script made using \link{alf_prep_slurm}.
This function does not need to be called directly if \link{alf_prep_slurm} is run with the default \code{copy_rscript = TRUE}.

This function is intended to be used in conjunction with \link{alf_prep_slurm}.
It creates a copy on disk at \code{out_dir} of a template R script contained in the \code{alfresco} package.
}
\examples{
\dontrun{alf_prep_rscript()}
}
