% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{alf_prep_slurm}
\alias{alf_prep_slurm}
\title{Slurm script for ALFRESCO probability distributions}
\usage{
alf_prep_slurm(out_dir = alfdef()$alf_slurm_dir, file = "alf_prep.R",
  inputs = NULL, outputs = NULL, project = NULL, period = NULL,
  reps = NULL, email = "mfleonawicz@alaska.edu", exclusive = TRUE,
  copy_rscript = TRUE)
}
\arguments{
\item{out_dir}{the directory containing \code{file}. This is where the slurm script is to be saved and executed from.}

\item{file}{the R script to be called by \code{Rscript} when the generated slurm script is executed.
May be similarly generated by this function. See \code{copy_rscript} below.}

\item{inputs}{override the default directory where input .rds files containing extracted data are stored.}

\item{outputs}{override the default directory where distributions .rds files are output.}

\item{project}{character, a (new) project name for extracted data. It need not match any directory names pertaining to the raw ALFRESCO outputs.}

\item{period}{character, \code{"historical"} or \code{"projected"}.}

\item{reps}{numeric vector of ALFRESCO simulation replicates for data extraction, e.g. \code{1:200}.}

\item{email}{defaults to the author/maintainer/user.}

\item{exclusive}{logical, put CPUs on an Atlas compute node into exclusive use for the job when the generated slurm script is executed. Defaults to \code{TRUE}.}

\item{copy_rscript}{logical, also copy template R script from \code{alfresco} package along with the generated slurm script. Defaults to \code{TRUE}.}
}
\description{
Generate slurm script for curating probability distributions from extracted ALFRESCO data.
}
\details{
This function is used for generating a slurm script that is used for curating probability distributions of random variables previously extracted
from ALFRESCO geotiff map outputs. The distributions are stored in data frames in .rds files for subsequent analyses in various projects.
The generated slurm script is used to call an R script with \code{Rscript}, e.g., that created by \link{alf_prep_rscript}.

Formal SLURM job arguments are passed by \code{exclusive} and \code{email}.
General script setup arguments include \code{out_dir}, \code{file} and \code{copy_rscript}.
All other arguments refer to those passed to the \code{Rscript} call within the slurm script.
These include \code{inputs}, \code{outputs}, \code{project}, \code{period} and \code{reps}.
If \code{NULL}, they are ignored.
Provide non-\code{NULL} arguments to override \code{alfresco} package defaults for \code{inputs} and \code{outputs}.
It is assumed that any of these four arguments will be passed as name-value pairs (in any order)
at the command line by the user when the generated slurm script is executed.
Any of these arguments that are not \code{NULL} are hardcoded into the string of arguments listed after \code{Rscript}.

This provides flexible generality when generating ALFRESCO data prep slurm scripts.
Note that the number of non-null arguments among these decreases the number of available general arguments available at the command line
when the script is executed. For example, if three of five arguments available to \code{Rscript} are hardcoded into the slurm
script by passing them to \code{alf_prep_slurm} explicitly, then the generated script will show only an additional \code{$1 $2}
after the fixed arguments rather than \code{$1 $2 $3 $4 $5}.

Note that this function is intended to be run on the Atlas cluster. If you make a bash script like this on Windows,
you may have to run a command line utility like \code{dos2unix} on the file.
}
\examples{
\dontrun{
alf_prep_slurm(project = "JFSP", reps = 1:32)
}

}
