% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfresco.R
\docType{package}
\name{alfresco}
\alias{alfresco}
\alias{alfresco-package}
\title{alfresco: R functions for ALFRESCO wildfire model.}
\description{
The alfresco package provides functions for working with and analyzing ALFRESCO wildfire model inputs and outputs.
It is a member package in the functional sector of the SNAPverse.
}
