% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmo.R
\name{ba_fmo}
\alias{ba_fmo}
\title{Burn area by FMO zone}
\usage{
ba_fmo(in_dir, years, id = 0:5, labels = c("Unmanaged", "Limited",
  "Modified", "Critical", "Full", "Other"))
}
\arguments{
\item{in_dir}{input directory, a \code{Maps} directory for ALFRESCO run output.}

\item{years}{integer.}

\item{id}{integer, 0 through 5 available, pertaining to ID codes for FMO zones.}

\item{labels}{character, labels for \code{id} values.}
}
\value{
a data frame.
}
\description{
Generate a data frame of mean (across replicates) annual burn area by fire management options zones.
}
\examples{
\dontrun{
ba_fmo(".", 1950:2013, 0:5)
}
}
