% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv.R
\name{cbdf}
\alias{cbdf}
\title{Cumulative burn over time data frame}
\usage{
cbdf(x)
}
\arguments{
\item{x}{character or data frame. The file name of the or the data frame object from that file. See details.}
}
\value{
a data frame.
}
\description{
Generate a data frame of cumulative burn over time for the union of all vegetation classes in a region.
}
\details{
This function takes an exact-type fire size by vegetation class .rds file or the data frame it contains if already loaded into the R session.
The source file is one created by \link{fsv}.
Vegetation-specific fire sizes are aggregated by unique fire IDs per simulation replicate, year and,
if present, fire management options treatment levels. It then adds a column for cumulative burn area
alongside ascending years.
}
\examples{
\dontrun{cbdf("historical_fsv.rds")}
}
\seealso{
fsv fsdf
}
