% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{copy_alf_outputs}
\alias{copy_alf_outputs}
\title{Generate bash script for copying ALFRESCO map outputs}
\usage{
copy_alf_outputs(project, in_dir, out_dir = alfdef()$alf_extract_dir,
  file = "copy_alf_outputs.sh", domain = "akcan1km", cru_id = "CRU32")
}
\arguments{
\item{project}{character, a (new) name for the project, not required to match anything in source data directories.}

\item{in_dir}{input directory, the parent directory above model (GCM/RCP) run subdirectories.}

\item{out_dir}{output directory.}

\item{file}{name of shell script.}

\item{domain}{character, the ALFRESCO run spatial domain, either \code{"akcan1km"} or \code{"ak1km"}.}

\item{cru_id}{character, label for CRU data. Defaults to \code{"CRU 3.2"}.}
}
\value{
invisible, writes a file to disk.
}
\description{
Generate bash script for copying ALFRESCO map outputs from one file system location to another.
}
\details{
This function is used for generating a bash script that is used for making a copy of a subset of ALFRESCO geotiff map outputs for subsequent analyses.
The historical and projected periods (years subset) are inferred from the choice of ALFRESCO spatial domain.
The file copy process adjusts the new file path structure, creates more consistency, and limits the number of geotiffs
to only to required ones. This means not only limiting to standard year ranges but also retaining only
fire scar, vegetation type and vegetation age maps. The bash script itself is run on SNAP's Atlas cluster.

Note that this function is intended to be run on the Atlas cluster. If you make a bash script like this on Windows,
you may have to run a command line utility like \code{dos2unix} on the file.
}
\examples{
\dontrun{
copy_alf_outputs(
  "JFSP",
  "/big_scratch/shiny/Runs_Statewide/JFSP",
  domain = "ak1km")
}
}
