% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_alf_stops}
\alias{extract_alf_stops}
\title{ALFRESCO extraction stop error checks}
\usage{
extract_alf_stops()
}
\value{
invisible, or an error.
}
\description{
Helper function to check for missing objects and stop process if necessary.
}
\examples{
\dontrun{extract_alf_stops}
}
