% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{fire_perimeters}
\alias{fire_perimeters}
\title{Obtain Alaska fire perimeter data}
\usage{
fire_perimeters(years = 1940:2017)
}
\arguments{
\item{years}{numeric vector, defaults to all available years, 1940 - 2017.}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Obtain Alaska Fire Service historical Alaska fire perimeter data.
}
\details{
The most current fire perimeter data available through this function is up through 2017.
The data is pulled from SNAP via Amazon Web Services. It is downloaded and imported from a ~21 MB rds file as a \code{SpatialPolygonsDataFrame}.
For slow connections, please allow time for download. Subsetting years does not affect this because the full file is imported and any subsetting is done subsequently.
}
\examples{
\dontrun{x <- fire_perimiters()}
}
