% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_weights.R
\name{fire_probs}
\alias{fire_probs}
\title{Fire point probabilities}
\usage{
fire_probs(data, veg_labels, covariates = TRUE)
}
\arguments{
\item{data}{a data frame output by \code{fire_weights}.}

\item{veg_labels}{character, vector of vegetation labels. If missing, they are assumed from \code{get_veg_labels("ak1km")}.}

\item{covariates}{logical, if \code{TRUE} (default), then a wide data frame is returned that includes additional spatially aggregated covariates. See details.}
}
\value{
a data frame.
}
\description{
Estimate fire point probability from fire weights for point locations.
}
\details{
This function estimates point probability of fire for spatial point locations from fire weights output by \link{fire_weights}.
It generates a summary data frame that includes a column \code{prop} of a simple proportion of area burned and a column \code{prob}
of estimated fire point probabilities based on the weights in \code{data}.

\code{covariates = TRUE} returns additional variables only if at least the \code{veg} column is present in \code{data}. If not, there will be
nothing additonal to return.

This function returns a one-column data frame. It is called on a single point location just like \code{fire_weights}.
It is generally used in a context where it is invoked multiple times applied to several point locations.
See \link{point_probs} for the generic wrapper used to estimate fire point probability for multiple locations that returns a multi-row data frame.
}
\examples{
\dontrun{
scar <- "FireScar_0_2004.tif"
veg <- "Veg_0_2004.tif"
age <- "Age_0_2004.tif"
coords <- data.frame(lon = -147.7164, lat = 64.8378)
fire_weights(r = scar, xy = coords, veg = veg, age = age)
fire_probs(x)
}
}
