% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_frp.R
\name{fire_scar_brick}
\alias{fire_scar_brick}
\title{Prepare fire scar bricks}
\usage{
fire_scar_brick(fire_area_history, fah_yrs, emp_yrs, emp_fire_cause,
  domain, shp, template_raster, mc.cores = 32)
}
\arguments{
\item{fire_area_history}{shapefile of fire area history.}

\item{fah_yrs}{fire area history years.}

\item{emp_yrs}{empirical data years.}

\item{emp_fire_cause}{character, cause of fire: \code{"lightning"} or \code{"all"}.}

\item{domain}{character, ALFRESCO domain.}

\item{shp}{masking shapefile.}

\item{template_raster}{raster layer.}

\item{mc.cores}{number of processors.}
}
\value{
side effect of reading and/or writing files and pushing objects to global environment.
}
\description{
Prepare raster bricks of empirical fire scar data.
}
\details{
This function loads relevant fire scar raster brick objects into the global environment from disk if they exist.
Otherwise it generates them from scratch, saves to disk for quicker use next time the function is called during a
future ALFRESCO run, and similarly pushes the objects to the global environment.
}
\examples{
# not run
}
