% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_weights.R
\name{fire_weights}
\alias{fire_weights}
\title{Fire probability weights for point locations}
\usage{
fire_weights(r, xy, buffer = 20000, weight = "linear",
  lonlat_to_akalbers = TRUE, veg = NULL, age = NULL)
}
\arguments{
\item{r}{raster layer or path to file, an ALFRESCO fire scar output geotiff.}

\item{xy}{a single location, in a 2-column data frame or matrix of longitude and latitude.}

\item{buffer}{radius of buffer around \code{xy}.}

\item{weight}{inverse distance weighting, \code{"linear"} or \code{"quadratic"}.}

\item{lonlat_to_akalbers}{if \code{xy} is unprojected (lon/lat), reproject to NAD83 Alaska Albers equal area conic projection.}

\item{veg}{same as\code{r} but for accompanying vegetation class ALFRESCO output.}

\item{age}{same as\code{r} but for accompanying vegetation age ALFRESCO output. See details.}
}
\value{
a data frame containing a logical column indicating burn or no burn,
a column of distances from point location, and a column of weights.
}
\description{
Compute fire probability weights for point locations.
}
\details{
This function applies an inverse distance weighting to grid cells within a radial buffer of a center grid cell.
The function is used for extracting values within a buffer around a given point from ALFRESCO fire scar output geotiffs.
Points outside the boundary defined by \code{buffer} are excluded and have zero
influence on estimated probability of fire at the center point.
Fire scars that intersect the radial buffer zone contribute more to the estimated likelihood of fire
for the point location the closer the fire is to the point.

It is recommended to not use \code{age} because ALFRESCO has a longstanding bug involving geographic areas of uninitialized ages,
resulting in extreme negative age values that ruin analyses unless the user first implements their own fix to any age geotiffs before using them.
}
\examples{
\dontrun{
scar <- "FireScar_0_2004.tif"
veg <- "Veg_0_2004.tif"
age <- "Age_0_2004.tif"
coords <- data.frame(lon = -147.7164, lat = 64.8378)
fire_weights(r = scar, xy = coords, veg = veg, age = age)
}
}
