% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{fireEventsFun}
\alias{fireEventsFun}
\alias{fireEventsFunEmpirical}
\title{ALFRESCO post-processing of fire events}
\usage{
fireEventsFun(k, pts, locs, replicates, source = "Modeled",
  buffer.list = list(NULL), buffer.labels = LETTERS[1:length(buffer.list)],
  burnable_cells = NULL, main_dir)

fireEventsFunEmpirical(b, pts, locs, replicates = "Observed",
  source = "Observed", buffer.list = list(NULL),
  buffer.labels = LETTERS[1:length(buffer.list)], burnable_cells = NULL)
}
\arguments{
\item{k}{integer, an iterator.}

\item{pts}{table of point locations.}

\item{locs}{location names}

\item{replicates}{character, simulation replicate labels.}

\item{source}{character, \code{"Observed"} or \code{"Modeled"}.}

\item{buffer.list}{list of point location buffers in meters.}

\item{buffer.labels}{labels for buffers in list.}

\item{burnable_cells}{binary raster of possibly burnable cells.}

\item{main_dir}{directory path.}

\item{b}{a raster brick.}
}
\value{
a list.
}
\description{
These functions are used for post-processing of ALFRESCO fire events from map outputs.
}
