% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmo.R
\name{fmo_add_buffer}
\alias{fmo_add_buffer}
\title{Add buffer to FMO zone base map}
\usage{
fmo_add_buffer(out_dir = ".", file = "fmo_buffer.tif",
  unmanaged = NULL, limited = NULL, modified = NULL, full = NULL,
  critical = NULL, other = NULL)
}
\arguments{
\item{out_dir}{output directory.}

\item{file}{output filename.}

\item{unmanaged}{numeric.}

\item{limited}{numeric.}

\item{modified}{numeric.}

\item{full}{numeric.}

\item{critical}{numeric.}

\item{other}{numeric.}
}
\value{
nothing is returned but a file is written to disk.
}
\description{
Add a buffer around any FMO zone in the FMO zone raster base map.
}
\details{
This function allows for adding a buffer of different radii (in meters) around any of the unique FMO zones.
}
\examples{
\dontrun{fmo_add_buff(full = 5000)}
}
