% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmo.R
\name{fmo_cb_reduction}
\alias{fmo_cb_reduction}
\title{FMO cumulative burn reduction table}
\usage{
fmo_cb_reduction(data)
}
\arguments{
\item{data}{a data frame resulting from \link{fsdf}.}
}
\value{
a data frame.
}
\description{
Generate a data frame of reductions in cumulative burn based on fire management options treatment levels.
}
\examples{
\dontrun{
x <- fsdf("historical_fsv.rds")
fmo_cb_reduction(x)
}
}
