% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv.R
\name{fsv}
\alias{fsv}
\title{Compute fire size by vegetation class}
\usage{
fsv(in_dir, out_dir, group, region, period)
}
\arguments{
\item{in_dir}{input \code{fsv} directory where extracted \code{fsv} data are located in .rds files.}

\item{out_dir}{top level (project) output directory: the function will add a \code{fsv/[group]/[region]} subdirectory where the output file is saved.}

\item{group}{character, the region group, used for file selection in \code{in_dir}.}

\item{region}{character, the region, used for file selection in \code{in_dir}.}

\item{period}{character, \code{"historical"} or \code{"projected"}.}
}
\description{
Compute and save to disk exact fire size by vegetation class from extracted ALFRESCO outputs.
}
\details{
This function computes exact annual fire sizes by vegetation from extracted ALFRESCO
model runs. If discrete probability distributions of annual fire by vegetation class are required, see \link{alf_dist}.
This function uses the same inputs, the extraction .rds files created by \link{run_alf_extraction}.
It essentially does nothing more than row bind the data frames from the extraction files without computing anything other statistics.
This yields one file in a project directory that covers all extracted fire size by vegetation class data for a unique geographic region.
}
\examples{
\dontrun{
in_dir <- "/atlas_scratch/mfleonawicz/alfresco/JFSP/extractions/fsv"
out_dir <- "/workspace/UA/mfleonawicz/data/alfresco/JFSP"
fsv(in_dir, out_dir, "FMO", "MFC buffers", "historical")
}
}
