% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_domain_cells}
\alias{get_domain_cells}
\title{Load cell index data frame for subregions}
\usage{
get_domain_cells(domain = "akcan1km", cells)
}
\arguments{
\item{domain}{character, the ALFRESCO run spatial domain, either \code{"akcan1km"} or \code{"ak1km"}.}

\item{cells}{character, file name of cell index table rds file. If missing, default path taken from \code{snapprep::snapdef()} based on \code{domain}.}
}
\value{
a data frame.
}
\description{
Load data frame of raster grid cell indices defining groups of spatial subregions.
}
\examples{
\dontrun{cells <- get_domain_cells()}
}
