% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_out_dirs}
\alias{get_out_dirs}
\title{Get ALFRESCO output directories}
\usage{
get_out_dirs(domain, project, cru)
}
\arguments{
\item{domain}{character, the ALFRESCO run spatial domain, either \code{"akcan1km"} or \code{"ak1km"}.}

\item{project}{character, valid projects based on the domain. See details.}

\item{cru}{logical, whether data extraction is for historical years (ALFRESCO runs based on CRU data) or projected years (GCM data).}
}
\value{
a vector of project climate model/emissions scenario ALFRESCO output directories.
}
\description{
Get ALFRESCO output directories for an ALFRESCO project.
}
\details{
This function supports data extraction procedures for ALFRESCO model geotiff map outputs.
A project refers to a collection of output directories pertaining to the set of climate models and emissions scenarios/RCPs
that were used in project simulations. This function returns the full file path to each pertinent directory, given a valid \code{domain}
and \code{project}.
Valid projects for Alaska/western Canada include \code{"IEM"} and \code{"FMO_Calibrated"}. For Alaska "statewide", it is \code{"CMIP5_SW"}.
}
\examples{
\dontrun{domain = "ak1km", project = "CMIP5_SW", cru = TRUE}
}
