% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_veg_labels}
\alias{get_veg_labels}
\title{Get vegetation labels}
\usage{
get_veg_labels(domain = "akcan1km")
}
\arguments{
\item{domain}{character, the ALFRESCO run spatial domain, either \code{"akcan1km"} or \code{"ak1km"}.}
}
\value{
a vector of vegetation labels.
}
\description{
Get vegetation labels based on ALFRESCO domain, which is suggestive of the vegetation base map.
}
\examples{
get_veg_labels()
}
