% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_fmo_ba}
\alias{obs_fmo_ba}
\title{Historical Alaska annual burn area by fire management options zones (FMOs).}
\format{A data frame with 64 rows and 8 columns: Set (Observed), Year, and six columns providing the total burn area within each of six FMOs: Unmanaged, Limited, Modified, Full, Critical and Other.
There is one row for each year.}
\usage{
obs_fmo_ba
}
\description{
A data frame of annual empirical, observed historical burn areas in Alaska assembled from 1950 - 2013 fire perimeter data.
The total burn area is segmented into that falling into each of six exhaustive and mutually exclusive regional zones based on existing fire management options.
}
\keyword{datasets}
