% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_as_inputs.R
\name{outputs_as_inputs}
\alias{outputs_as_inputs}
\title{Copy ALFRESCO output maps}
\usage{
outputs_as_inputs(domain, run_name, year, suffix = "runs",
  sub_dir = "secondaryRunInputs", mc.cores = 32)
}
\arguments{
\item{domain}{character, usually \code{"Statewide"} or \code{"Noatak"}.}

\item{run_name}{character.}

\item{year}{numeric, the "final" year of a run. See details.}

\item{suffix}{appended to output directory, defaults to \code{"runs"}.}

\item{sub_dir}{name of directory where copied maps are stored, defaults to \code{"secondaryRunInputs"}.}

\item{mc.cores}{number of processors, defaults to 32.}
}
\value{
file copy side effect.
}
\description{
Copy ALFRESCO output maps to new location for use as inputs to subsequent model runs.
}
\details{
This function copies ALFRESCO output maps from a single, "final" year where \code{year} is chosen to be
one less than the starting year of the subsequent model run. For example, a run beginning in 1950
would use as inputs 1949 output maps from a prior model run.
\code{year} need not be the final year of the original run, but it is final
in the context that the following year will be part of a new, subsequent run.
The output directory is at alf_defaults()$atlas_shiny_dir and begins with \code{"/Final_"}.
}
\examples{
\dontrun{
outputs_as_inputs("Statewide", "fmo00s00i_historical_CRU32", 1949)
outputs_as_inputs("Statewide", "fmo00s00i_historical_CRU32", 2013)
}
}
