% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_prob.R
\name{plot_map_zoom}
\alias{plot_map_zoom}
\title{Plot map composition for radial buffer around a point}
\usage{
plot_map_zoom(file, x, cells, radius, classes, col, title = NULL,
  class_order = seq_along(classes), base_path = ".", width = 1000,
  height = 1000)
}
\arguments{
\item{file}{character, output file.}

\item{x}{raster layer. Should be discretely valued/categorical.}

\item{cells}{cell numbers, typically defining a radial clump, but will accept any valid cells numbers.}

\item{radius}{in meters, catering to NAD83 Alaska Albers projected ALFRESCO map output.}

\item{classes}{character, the names of the classes assigned to the unique values in \code{x}.}

\item{col}{color vector the same length as \code{lasses} and the number of unique non-\code{NA} values in \code{x}.}

\item{title}{character, plot title.}

\item{class_order}{integer vector, optional reordering of classes when the ordered values in \code{x} are not the desired order.}

\item{base_path}{character, defaults to working directory.}

\item{width}{numeric, plot width in pixels.}

\item{height}{numeric, plot height in pixels.}
}
\value{
nothing is returned, but a file is saved to disk.
}
\description{
Plot the map composition of a raster layer in a radial neighborhood of a point.
}
\details{
This function saves a png to disk of a radial buffer around a point. It is used for zooming in spatially around points of interest in raster layers such as those output by ALFRESCO.
}
\examples{
# not run
}
