% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_prob.R
\name{point_fire}
\alias{point_fire}
\title{Probability of fire}
\usage{
point_fire(rep, base_path = ".", label, center, sample_size = 100,
  sample_buffer = 25000, extraction_buffers = seq(1000, 30000, length.out =
  100), years)
}
\arguments{
\item{rep}{integer, ALFRESCO replicate.}

\item{base_path}{input directory to ALFRESCO fire scar geotiffs.}

\item{label}{character, name to attach to raster layer sum for input fire scar geotiffs.}

\item{center}{numeric, coordinate pair for centroid around which buffered extractions are taken.}

\item{sample_size}{integer, number of grid cells to sample within \code{sample_buffer} of \code{center}.}

\item{sample_buffer}{numeric, buffer radius in meters around \code{center} to sample grid cells.}

\item{extraction_buffers}{numeric, vector of buffer radii in meters from each sampled point to compute the proportion of cells where fire has occurred.}

\item{years}{numeric, vector of years for which to compute the probability of fire.}
}
\value{
a data frame.
}
\description{
Compute the probability of fire within some radius of a point.
}
\details{
This function computes the proportion of cells within a radial buffer zone of a sampled grid cell that has burned in ALFRESCO fire scar geotiff outputs.
It does this for each of a number of sampled points around a centroid point of interest. There are radius arguments for both the size of the buffer around the center from
which to draw random grid cell samples for bootstrapping a more robust estimate around the centroid grid cell as well as the buffer around each sampled point within
which to compute the likelihood of fire. The latter buffer argument may be a vector. This function sums fire occurrences across replicates and years.
}
\examples{
# not run
}
