% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_weights.R
\name{point_probs}
\alias{point_probs}
\title{Fire probabilities for point locations}
\usage{
point_probs(r, xy, id, buffer = 20000, weight = "linear",
  lonlat_to_akalbers = TRUE, veg_labels, covariates = TRUE,
  veg = NULL, age = NULL, max.cores = 32)
}
\arguments{
\item{r}{raster layer or path to file, an ALFRESCO fire scar output geotiff.}

\item{xy}{a 2-column data frame or matrix of longitude and latitude.}

\item{id}{a vector of IDs or names associated with the locations in the rows of \code{xy}.}

\item{buffer}{radius of buffer around \code{xy}.}

\item{weight}{inverse distance weighting, \code{"linear"} or \code{"quadratic"}.}

\item{lonlat_to_akalbers}{if \code{xy} is unprojected (lon/lat), reproject to NAD83 Alaska Albers equal area conic projection.}

\item{veg_labels}{character, vector of vegetation labels. If missing, they are assumed from \code{get_veg_labels("ak1km")}.}

\item{covariates}{logical, if \code{TRUE} (default), then a wide data frame is returned that includes additional spatially aggregated covariates. See details.}

\item{veg}{same as\code{r} but for accompanying vegetation class ALFRESCO output.}

\item{age}{same as\code{r} but for accompanying vegetation age ALFRESCO output. See details.}

\item{max.cores}{integer, maximum number of processors, defaults to 32 for Atlas compute node.}
}
\value{
a data frame containing a logical column indicating burn or no burn,
a column of distances from point location, and a column of weights.
}
\description{
Estimate fire probability for multiple point locations.
}
\details{
This function is a wrapper around \link{fire_weights} and \link{fire_probs}.
It uses parallel processing via \code{parallel::mclapply} to apply these functions to multiple point locations.
See these functions for further details.

It is recommended to not use \code{age} because ALFRESCO has a longstanding bug involving geographic areas of uninitialized ages,
resulting in extreme negative age values that ruin analyses unless the user first implements their own fix to any age geotiffs before using them.
}
\examples{
\dontrun{
scar <- "FireScar_0_2004.tif"
veg <- "Veg_0_2004.tif"
age <- "Age_0_2004.tif"
coords <- data.frame(lon = c(-146, -150), lat = c(64, 67))
point_probs(r = scar, xy = coords, id = c("point1", "point2"), veg = veg, age = age)
}
}
