% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_frp.R
\name{pp_df_prep}
\alias{pp_df_prep}
\title{Prepare data frame during ALFRESCO post-processing}
\usage{
pp_df_prep(x, multiple_reps = TRUE)
}
\arguments{
\item{x}{output from \code{fire_events} or \code{fire_events_emp}.}

\item{multiple_reps}{logical, typically \code{TRUE} for ALFRESCO model output and \code{FALSE} for empirical data.}
}
\value{
a tibble.
}
\description{
Prepare data frame during post-processing of ALFRESCO fire events from map outputs.
}
\examples{
# not run
}
