% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_frp.R
\name{pp_fire_events}
\alias{pp_fire_events}
\title{Prepare list of data frames during ALFRESCO post-processing}
\usage{
pp_fire_events(alf_data, emp_data, domain, group_name, run_name, emp_yrs, out,
  send_to_app = TRUE)
}
\arguments{
\item{alf_data}{output from \code{fire_events}.}

\item{emp_data}{output from \code{fire_events_emp}.}

\item{domain}{character, ALFRESCO domain.}

\item{group_name}{character, ALFRESCO run group name.}

\item{run_name}{character, unique ALFRESCO run name.}

\item{emp_yrs}{empirical years.}

\item{out}{character, output directory.}

\item{send_to_app}{logical, send processed fire event outputs to Shiny app on Eris server. Defaults to \code{TRUE}.}
}
\value{
side effect of reading and writing files.
}
\description{
Prepare list of data frames during post-processing of ALFRESCO fire events from map outputs.
}
\examples{
# not run
}
