% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_fsv.R
\name{pp_fsv}
\alias{pp_fsv}
\alias{pp_fsv_emp}
\alias{pp_fsv_save}
\title{Fire size by vegetation ALFRESCO post-processing}
\usage{
pp_fsv(d, main_dir, veg, shp, years, obs = NULL)

pp_fsv_emp(i, b, veg, shp, years, obs = NULL)

pp_fsv_save(d, main_dir, veg, shp, alf_yrs, emp_yrs, obs, i, b, out, domain,
  mc.cores = 32)
}
\arguments{
\item{d}{iterator, ALFRESCO replicates.}

\item{main_dir}{input directory.}

\item{veg}{vegetation raster layer.}

\item{shp}{masking shapefile.}

\item{years}{numeric, vector of years.}

\item{obs}{fire scar period totals raster layer.}

\item{i}{iterator, layer number of \code{b}.}

\item{b}{annual fire ID raster brick.}

\item{alf_yrs}{ALFRESCO run years.}

\item{emp_yrs}{empirical data years.}

\item{out}{character, output directory.}

\item{domain}{character, ALFRESCO domain, \code{Statewide} or \code{Noatak}.}

\item{mc.cores}{number of processors.}
}
\value{
a tibble.
}
\description{
Annual fire size by vegetation class ALFRESCO post-processing. \code{pp_fsv} works with model outputs.
The companion function, \code{pp_fsv_emp} handles empirical data.
\code{pp_fsv_save} is a wrapper function that calls both and also generates and saves plots and data to disk.
}
\examples{
# not run
}
