% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandline.R
\name{prep_alf_frp}
\alias{prep_alf_frp}
\title{Helper function for FRP processing}
\usage{
prep_alf_frp(comArgs, out)
}
\arguments{
\item{out}{}
}
\value{
side effects of creating output directory, \code{sink}-ing message.
}
\description{
Helper function for Fire Return Period processing.
}
\examples{
# not run
}
