% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandline.R
\name{prep_comArgs}
\alias{prep_comArgs}
\title{Prepare command line arguments}
\usage{
prep_comArgs(comArgs)
}
\arguments{
\item{comArgs}{output from \code{commandArgs(TRUE)}.}
}
\value{
a matrix.
}
\description{
Prepare command line argument inputs in a matrix.
}
\examples{
# not run
}
