% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_frp.R
\name{prep_fire_events}
\alias{prep_fire_events}
\alias{prep_fire_events_emp}
\title{ALFRESCO post-processing of fire events}
\usage{
prep_fire_events(k, pts, locs, replicates, source = "Modeled",
  buffer_list = list(NULL), buffer_labels = LETTERS[1:length(buffer_list)],
  veg = NULL, main_dir)

prep_fire_events_emp(b, pts, locs, replicates = "Observed",
  source = "Observed", buffer_list = list(NULL),
  buffer_labels = LETTERS[1:length(buffer_list)], veg = NULL, yrs)
}
\arguments{
\item{k}{integer, an iterator.}

\item{pts}{table of point locations.}

\item{locs}{location names}

\item{replicates}{character, simulation replicate labels.}

\item{source}{character, \code{"Observed"} or \code{"Modeled"}.}

\item{buffer_list}{list of point location buffers in meters.}

\item{buffer_labels}{labels for buffers in list.}

\item{veg}{vegetation raster layer.}

\item{main_dir}{directory path.}

\item{b}{a raster brick.}

\item{yrs}{vector of years.}
}
\value{
a list.
}
\description{
These functions are used for post-processing of ALFRESCO fire events from map outputs.
}
