% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_frp.R
\name{prep_points}
\alias{prep_points}
\title{Prepare points matrix and locations vector}
\usage{
prep_points(pts, path, domain)
}
\arguments{
\item{pts}{a lon/lat matrix.}

\item{path}{file path.}

\item{domain}{ALFRESCO domain.}
}
\value{
a list.
}
\description{
Prepare points matrix and locations vector for buffered FRP extractions
}
\details{
This function also makes adjustments for Noatak domain if applicable.
}
\examples{
# not run
}
