% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{run_alf_extraction}
\alias{run_alf_extraction}
\title{Run ALFRESCO data extraction}
\usage{
run_alf_extraction(domain = "akcan1km", type, loop_by = "rep", main_dir,
  out_dir, project, reps, years, cells, veg_labels, cru = FALSE,
  cru_id = "CRU 3.2", itervar = seq_along(years), mc.cores = 32,
  rmpi = TRUE)
}
\arguments{
\item{domain}{character, the ALFRESCO run spatial domain, either \code{"akcan1km"} or \code{"ak1km"}.}

\item{type}{\code{"fsv"} or \code{"av"}. See details.}

\item{loop_by}{\code{"rep"} or \code{"year"} (default).}

\item{main_dir}{input directory.}

\item{out_dir}{output directory. If missing, a default is provided by \code{alfdef()$alf_extract_dir}.}

\item{project}{character, valid projects based on the domain. See details.}

\item{reps}{integer vector, simulation replicates included in data extraction, e.g., \code{1:200}.}

\item{years}{ALFRESCO model run years included in data extraction.}

\item{cells}{data frame of raster grid cell indices appropriate to \code{domain}. See \code{get_domain_cells}.}

\item{veg_labels}{vegetation labels appropriate to domain/vegetation input map for ALFRESCO runs. See \code{get_veg_labels}.}

\item{cru}{logical, whether data extraction is for historical years (ALFRESCO runs based on CRU data) or projected years (GCM data).}

\item{cru_id}{character, label for CRU data. Defaults to \code{"CRU 3.2"}.}

\item{itervar}{integer vector, iterator, defaults to \code{1:length(years)}.}

\item{mc.cores}{number of processors. See details.}

\item{rmpi}{logical, use Rmpi. Defaults to \code{TRUE}.}
}
\value{
invisible, writes files.
}
\description{
Extract data for different variables from ALFRESCO model outputs.
}
\details{
The extracted data depends on \code{type}, which is \code{"fsv"} for fire size by vegetation class data or
\code{"av"} for vegetation age and vegetation cover area data.
A project refers to a collection of output directories pertaining to the set of climate models and emissions scenarios/RCPs
that were used in project simulations. This function returns the full file path to each pertinent directory, given a valid \code{domain}
and \code{project}.
Valid projects for Alaska/western Canada include \code{"IEM"} and \code{"FMO_Calibrated"}. For Alaska "statewide", it is \code{"CMIP5_SW"}.
\code{mc.cores} is used explicitly when \code{rmpi = FALSE} for \code{parallel::mclapply} instead of multi-node processing.
}
\examples{
# Not run; decontextualized example.
\dontrun{run_alf_extraction(type = "fsv", main_dir = "Maps", project = "MyProject",
  reps = 1:200, years = 2014:2099, cells = cells, veg_labels = veg_labels)}
}
