% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmo.R
\name{save_fmo_panel}
\alias{save_fmo_panel}
\title{Save FMO panel plot}
\usage{
save_fmo_panel(out_dir = ".", width = 1200, height = 800)
}
\arguments{
\item{out_dir}{output directory.}

\item{width}{numeric, png width.}

\item{height}{numeric, png height.}
}
\description{
Save fire management options panel plot showing each FMO layer.
}
\details{
This functions save a png graphic of the 15-km buffered FMO management IDs raster layer from the \code{snapgrid} package.
}
\examples{
save_fmo_panel()
}
