% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{wgs2ak}
\alias{wgs2ak}
\title{Convert from WGS84 to Alaska Albers projection}
\usage{
wgs2ak(xy)
}
\arguments{
\item{xy}{a 2-column matrix or data frame of lon/lat coordinates.}
}
\value{
data frame with new attached coordinates.
}
\description{
Simple wrapper around \code{spTransform} to convert from WGS84 to Alaska Albers projection.
}
\examples{
# not run
}
