% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmo.R
\name{ba_fmo}
\alias{ba_fmo}
\title{Burn area by FMO zone}
\usage{
ba_fmo(in_dir, years, id = 0:5, labels = c("Unmanaged", "Limited",
  "Modified", "Critical", "Full", "Other"), fmo_layer = NULL,
  mask = NULL, mask_value = NULL)
}
\arguments{
\item{in_dir}{input directory, a \code{Maps} directory for ALFRESCO run output.}

\item{years}{integer.}

\item{id}{integer, 0 through 5 available, pertaining to ID codes for FMO zones.}

\item{labels}{character, labels for \code{id} values.}

\item{fmo_layer}{raster layer, optional. If \code{NULL}, the standard FMO base map from the snapgrid package is used.}

\item{mask}{character, optional mask: \code{NULL}, \code{"ecoreg"} for \code{snappoly::ecoreg} ecoregions, or \code{"fmz"} for \code{snappoly::fmz} Fire Management Zones. See details.}

\item{mask_value}{character, ecoregion or FMZ name to mask by, not applicable if \code{mask = NULL}, in which case the full statewide Alaska JFSP ALFRESCO domain. See details.}
}
\value{
a data frame.
}
\description{
Generate a data frame of mean (across replicates) annual burn area by fire management options zones.
}
\details{
Options for Alaska ecoregion and Alaska fire management zone masking include \code{"Arctic Tundra"}, \code{"Bering Taiga"}, \code{"Bering Tundra"} and \code{"Intermontane Boreal"} and
eight FMZs ("DAS", "FAS", "GAD", "MID", "SWS", "TAD", "TAS", "UYD"), respectively.
The limited sets represent their spatial intersections with the JFSP ALFRESCO spatial domain.
}
\examples{
\dontrun{
ba_fmo(".", 1950:2013, 0:5)
}
}
