% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_weights.R
\name{fire_weights}
\alias{fire_weights}
\title{Fire probability weights for point locations}
\usage{
fire_weights(file, xy, buffer = 20000, weight = "linear",
  lonlat_to_akalbers = TRUE)
}
\arguments{
\item{file}{input file, an ALFRESCO fire scar output geotiff.}

\item{xy}{a single location, in a 2-column data frame or matrix of longitude and latitude.}

\item{buffer}{radius of buffer around \code{xy}.}

\item{weight}{inverse distance weighting, \code{"linear"} or \code{"quadratic"}.}

\item{lonlat_to_akalbers}{if \code{xy} is unprojected (lon/lat), reproject to NAD83 Alaska Albers equal area conic projection.}
}
\value{
a data frame containing a logical column indicating burn or no burn,
a column of distances from point location, and a column of weights.
}
\description{
Compute fire probability weights for point locations.
}
\details{
This function applies an inverse distance weighting to grid cells within a radial buffer of a center grid cell.
The function is used for extracting values within a buffer around a given point from ALFRESCO fire scar output geotiffs.
Points outside the boundary defined by \code{buffer} are excluded and have zero
influence on estimated probability of fire at the center point.
Fire scars that interset the radial buffer zone contribute more to the estimated likelihood of fire
for the point location the closer the fire is to the point.
}
\examples{
\dontrun{
coords <- data.frame(lon = -147.7164, lat = 64.8378)
fire_weights("someFile.tif", coords)
}
}
