% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv.R
\name{fsdf}
\alias{fsdf}
\title{Fire size and cumulative burn data frame}
\usage{
fsdf(x)
}
\arguments{
\item{x}{character or data frame. The file name of the or the data frame object from that file. See details.}
}
\value{
a data frame.
}
\description{
Generate a data frame of exact fire sizes and cumulative burn for the union of all vegetation classes in a region.
}
\details{
This function takes an exact-type fire size by vegetation class .rds file or the data frame it contains if already loaded into the R session.
The source file is one created by \link{fsv}.
Vegetation-specific fire sizes are aggregated by unique fire IDs per simulation replicate, year and,
if present, fire management options treatment levels. It also adds a column for cumulative burn area
alongside ascending fire sizes.
}
\examples{
\dontrun{fsdf("historical_fsv.rds")}
}
\seealso{
fsv cbdf
}
