% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandline.R
\name{prep_alf_frp}
\alias{prep_alf_frp}
\title{Helper function for FRP processing}
\usage{
prep_alf_frp(comArgs, out)
}
\arguments{
\item{comArgs}{output from \code{commandArgs(TRUE)}.}

\item{out}{character, output directory.}
}
\value{
invisible, side effects of creating output directory and \code{sink} message to disk.
}
\description{
Helper function for Fire Return Period processing.
}
\examples{
# not run
}
