% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{wgs2ak}
\alias{wgs2ak}
\title{Convert from Alaska Albers projection to WGS84}
\usage{
wgs2ak(xy)
}
\arguments{
\item{xy}{}
}
\value{
data frame with new attached coordinates.
}
\description{
Simple wrapper around \code{spTransform} to convert from Alaska Albers projection to WGS84.
}
\examples{
# not run
}
